requirements.txt
# Core RAG dependencies
sentence-transformers==2.2.2
faiss-cpu==1.7.4
gpt4all==2.8.2
langchain==0.1.17
langchain-community==0.0.37

# Document processing
pypdf2==3.0.1
python-docx==1.1.0
requests==2.31.0

# RAG evaluation
ragas==0.1.9
datasets==2.14.5

# Utilities
numpy==1.24.3
pandas==2.0.3
tqdm==4.66.1
python-dotenv==1.0.0

# Optional: for better text processing
beautifulsoup4==4.12.2